+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [CSE] Added *cse.resource.cnt.enableLimits* configuration. 
 - [CSE] Added "T" command to the console (display only sub-tree of the resource tree).
+- [CSE] Improved startup error messages when running in headless mode.
 
 ### Changed
 - [CSE] Default limits for &lt;container> are now disabled by default.